test(msteams): split the stream controller's presentation cases into their own suite
